Add basic HUD objects

+++ src/engine/core/input/InputManager.hpp
@@ -1,6 +1,8 @@
 #pragma once
 
 #include <SFML/Graphics.hpp>
+#include <string>
+#include <map>
 
 namespace Gengine {
     class InputManager {
@@ -11,7 +13,7 @@ namespace Gengine {
         bool isElemClicked(sf:: Sprite object, sf::Mouse::Button button, sf::RenderWindow* window);
         bool isElemClicked(sf::RectangleShape object, sf::Mouse::Button button, sf::RenderWindow* window);
         void initKeys();
-        sf::Vector2i getMousePosition(sf::RenderWindow* window);
+        sf::Vector2f getMousePosition(sf::RenderWindow* window);
         std::map<std::string, int> keyBinds;
     
     };
